feat: add self custody available field v4 get wallet
