+++ services/wallet/controllers_v4_test.go
@@ -7,6 +7,7 @@ import (
 	"context"
 	"crypto"
 	"crypto/ed25519"
+	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -14,8 +15,10 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	appctx "github.com/brave-intl/bat-go/libs/context"
 	errorutils "github.com/brave-intl/bat-go/libs/errors"
 	"github.com/brave-intl/bat-go/libs/middleware"
+	"github.com/brave-intl/bat-go/services/wallet/storage"
 
 	"github.com/brave-intl/bat-go/libs/clients"
 
@@ -556,6 +559,51 @@ func (suite *WalletControllersV4TestSuite) TestUpdateBraveWalletV4_ReputationCal
 	server.Handler.ServeHTTP(rw, request)
 
 	suite.Require().Equal(http.StatusInternalServerError, rw.Code)
+}
+
+func (suite *WalletControllersTestSuite) TestGetWalletV4() {
+	pg, _, err := wallet.NewPostgres()
+	suite.Require().NoError(err)
+
+	pub, _, err := ed25519.GenerateKey(nil)
+	suite.Require().NoError(err)
+
+	paymentID := uuid.NewV4()
+	w := &walletutils.Info{
+		ID:          paymentID.String(),
+		Provider:    "brave",
+		PublicKey:   hex.EncodeToString(pub),
+		AltCurrency: ptrTo(altcurrency.BAT),
+	}
+	err = pg.InsertWallet(context.TODO(), w)
+	suite.Require().NoError(err)
+
+	whitelistWallet(suite.T(), pg, w.ID)
+
+	allowList := storage.NewAllowList()
+
+	service, _ := wallet.InitService(pg, nil, nil, allowList, nil, nil, nil, nil, nil, nil, wallet.DAppConfig{})
+
+	handler := handlers.AppHandler(wallet.GetWalletV4(service))
+
+	req, err := http.NewRequest("GET", "/v4/wallets/"+paymentID.String(), nil)
+	suite.Require().NoError(err, "a request should be created")
+
+	req = req.WithContext(context.WithValue(req.Context(), appctx.NoUnlinkPriorToDurationCTXKey, "-P1D"))
+	rctx := chi.NewRouteContext()
+	rctx.URLParams.Add("paymentID", paymentID.String())
+	req = req.WithContext(context.WithValue(req.Context(), chi.RouteCtxKey, rctx))
+
+	rr := httptest.NewRecorder()
+	handler.ServeHTTP(rr, req)
+
+	suite.Assert().Equal(http.StatusOK, rr.Code)
+
+	var resp wallet.ResponseV4
+	err = json.Unmarshal(rr.Body.Bytes(), &resp)
+	suite.Require().NoError(err)
+
+	suite.Assert().Equal(true, resp.SelfCustodyAvailable["solana"])
 }
 
 func signUpdateRequest(req *http.Request, paymentID string, privateKey ed25519.PrivateKey) error {
